Usage-Rights-To-Metadata-Processor: move logic to common-lib and implement supplier processor
